requirements.txt
# Python K-Means Dataset Generation Dependencies
numpy
scikit-learn
cuml
cupy-cuda12x